+++ init.sh
@@ -1,5 +1,10 @@
+if [ "$EUID" -ne 0 ]; then
+    echo "Please run this script with sudo."
+    exit 1
+fi
+
 echo "[INIT]: UMOUNT boot.img..."
-sudo umount mounted/
+umount mounted/
 echo "[INIT]: RM boot.img..."
 rm boot.img
 echo "[INIT]: CREATE NEW boot.img..."
@@ -8,8 +13,8 @@ chmod 777 boot.img
 echo "[INIT]: CREATE NEW mounted dir"
 mkdir mounted
 echo "[INIT]: CREATE FILESYSTEM..."
-sudo mkfs boot.img
+mkfs boot.img
 echo "[INIT]: MOUNT boot.img TO mounted/ ..."
-sudo mount boot.img mounted/
+mount boot.img mounted/
 echo "[INIT]: INSTALLING SYSLINUX BOOTLOADER..."
-sudo extlinux --install mounted/
+extlinux --install mounted/
